+++ Debug/Core/Src/main.su
@@ -27,6 +27,16 @@ stm32f3xx_ll_bus.h:851:22:LL_APB2_GRP1_EnableClock	24	static
 stm32f3xx_ll_system.h:1618:22:LL_FLASH_SetLatency	16	static
 stm32f3xx_ll_system.h:1631:26:LL_FLASH_GetLatency	4	static
 stm32f3xx_ll_pwr.h:244:22:LL_PWR_EnableBkUpAccess	4	static
-main.c:68:5:main	48	static
-main.c:154:6:SystemClock_Config	8	static
-main.c:215:6:Error_Handler	4	static
+main.c:86:5:main	40	static
+main.c:181:6:SystemClock_Config	8	static
+main.c:235:6:estado_Anterior	16	static
+main.c:244:6:EDO_0	32	static
+main.c:275:6:EDO_1	32	static
+main.c:375:6:EDO_2	40	static
+main.c:405:6:EDO_3	32	static
+main.c:528:6:EDO_4	24	static
+main.c:544:6:EDO_5	32	static
+main.c:576:6:EDO_6	32	static
+main.c:587:6:EDO_7	32	static
+main.c:634:6:fsm	32	static
+main.c:645:6:Error_Handler	4	static
